python-backend/requirements.txt
# Agent7 Advanced Backend Dependencies

# Core
websockets>=12.0
asyncio-mqtt>=0.16.0
python-socketio>=5.11.0

# Browser Automation - Full browser-use capabilities
browser-use>=0.1.0
playwright>=1.40.0
selenium>=4.15.0

# LLM Providers
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
langchain-google-genai>=0.0.6
langchain-aws>=0.2.0
openai>=1.6.0
anthropic>=0.8.0
google-generativeai>=0.3.0
boto3>=1.34.0

# Vision & Image Processing
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
pytesseract>=0.3.10

# Data & Storage
pydantic>=2.5.0
dataclasses-json>=0.6.0
python-dotenv>=1.0.0

# Utilities
requests>=2.31.0
aiohttp>=3.9.0
httpx>=0.25.0

# macOS Integration
pyobjc>=10.0; sys_platform == 'darwin'
pyobjc-framework-Cocoa>=10.0; sys_platform == 'darwin'
pyobjc-framework-Quartz>=10.0; sys_platform == 'darwin'

# Advanced Features
tiktoken>=0.5.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
markdownify>=0.11.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
